+++ vite.config.ts
@@ -3,7 +3,7 @@ import react from '@vitejs/plugin-react'
 import { VitePWA } from 'vite-plugin-pwa'
 
 export default defineConfig({
-  define: { __APP_VERSION__: JSON.stringify('1.2.2') },
+  define: { __APP_VERSION__: JSON.stringify('1.3.0') },
   base: './',
   plugins: [react(), VitePWA({
     registerType: 'prompt',